// src/features/co-pilot/core/read-kind.types.ts
export type ReadQueryKind =
  | "NONE"
  | "ESCROW_AUDIT"
  | "ESCROW_TDS"
  | "ESCROW_SETUP"
  | "COLLAB_PIPELINE"
  | "COLLAB_ISSUES"
  | "DNA_COMPLIANCE"
  | "BRAND_CENTRE_GREETING"
  | "BRAND_CENTRE_OVERVIEW"
  | "BRAND_CENTRE_COMPLETENESS"
  | "BRAND_CENTRE_READINESS"
  | "BRAND_CENTRE_DNA_BLOCKS"
  | "BRAND_CENTRE_LEAKS"
  | "BRAND_CENTRE_PERSONAS"
  | "BRAND_CENTRE_COMPETITOR_INSIGHTS"
  | "BRAND_CENTRE_VISUAL_IDENTITY"
  | "PLANNER_PIPELINE"
  | "CAMPAIGN_DRAFT_LIST"
  | "LIST_CAMPAIGNS"
  | "SEARCH_CAMPAIGNS"
  | "FILTER_CAMPAIGNS"
  | "SORT_CAMPAIGNS"
  | "CAMPAIGN_SUMMARY"
  | "CAMPAIGN_PERFORMANCE"
  | "COMPARE_CAMPAIGNS"
  | "CAMPAIGN_FINANCIALS"
  | "BRAND_CENTRE_DEFAULT";